// src/entities/current-pipe-parameters/item/api/dto/create-sbt.dto.ts
import {
    ISbtDTOCoupling,
    ISbtDTONipple,
    ISbtDTOPipeBody
} from './sbt.dto'
import { CreateMainInfoPipeDTO } from './base-type/create-pipe.dto'

type OverrideCreateSbtPipeBodyDTO = 'average_wall_thickness' | 'total_length'

export interface CreateSbtPipeBodyDTO
    extends Omit<
        ISbtDTOPipeBody,
        'pipe_body_class' | OverrideCreateSbtPipeBodyDTO
    > {
    average_wall_thickness: number
    total_length: number
}

type OverrideCreateSbtNippleDTO =
    | 'outer_diameter_nipple'
    | 'chamfer_diameter_nipple'
    | 'nipple_thread_length'
    | 'key_installation_location_nipple'
    | 'inner_diameter_nipple'

export interface CreateSbtNippleDTO
    extends Omit<
        ISbtDTONipple,
        | 'condition_lock_connection_nipple'
        | 'lock_connection_class_nipple'
        | OverrideCreateSbtNippleDTO
    > {
    outer_diameter_nipple: number
    chamfer_diameter_nipple: number
    nipple_thread_length?: number
    key_installation_location_nipple: number
    inner_diameter_nipple: number
}

type OverrideCreateSbtCouplingDTO =
    | 'outer_diameter_coupling'
    | 'chamfer_diameter_coupling'
    | 'width_thrust_shoulder_coupling'
    | 'diameter_cone_recess_coupling'
    | 'depth_cone_recess_coupling'
    | 'key_installation_location_coupling'

export interface CreateSbtCouplingDTO
    extends Omit<
        ISbtDTOCoupling,
        | 'condition_lock_connection_coupling'
        | 'lock_connection_class_coupling'
        | OverrideCreateSbtCouplingDTO
    > {
    outer_diameter_coupling: number
    chamfer_diameter_coupling: number
    width_thrust_shoulder_coupling: number
    diameter_cone_recess_coupling: number
    depth_cone_recess_coupling: number
    key_installation_location_coupling: number
}

export type CreateSbtDTO = CreateSbtPipeBodyDTO &
    CreateSbtNippleDTO &
    CreateSbtCouplingDTO &
    CreateMainInfoPipeDTO & {
    minimum_wall_thickness_ultra: number
    minimum_wall_thickness_premium: number
    minimum_wall_thickness_class_2: number
}